requirements.txt
faker==24.0.0
numpy>=1.24
pandas>=2.0
psycopg2-binary>=2.9
scipy>=1.10
python-dotenv>=1.0
tqdm>=4.65.0
